Integrated SRA support through HTSJDK
