+++ public/gatk-engine/src/main/java/org/broadinstitute/gatk/engine/CommandLineExecutable.java
@@ -37,6 +37,7 @@ import org.broadinstitute.gatk.engine.io.stubs.VCFWriterArgumentTypeDescriptor;
 import org.broadinstitute.gatk.utils.refdata.utils.RMDTriplet;
 import org.broadinstitute.gatk.engine.walkers.Walker;
 import org.broadinstitute.gatk.utils.text.ListFileUtils;
+import htsjdk.samtools.sra.SRAAccession;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -66,6 +67,10 @@ public abstract class CommandLineExecutable extends CommandLineProgram {
     private final Collection<Object> argumentSources = new ArrayList<Object>();
 
     protected static Logger logger = Logger.getLogger(CommandLineExecutable.class);
+
+    static {
+        SRAAccession.setAppVersionString("GATK " + getVersionNumber());
+    }
 
     /**
      * this is the function that the inheriting class can expect to have called
